+++ src/HotelChatbot.Infrastructure/Class1.cs
@@ -1,6 +1,0 @@
-﻿namespace HotelChatbot.Infrastructure;
-
-public class Class1
-{
-
-}